remove nette
