--- newscoop/library/Nette/Reflection/Annotation.php
@@ -1,44 +1,0 @@
-<?php
-
-/**
- * This file is part of the Nette Framework (http://nette.org)
- *
- * Copyright (c) 2004, 2011 David Grudl (http://davidgrudl.com)
- *
- * For the full copyright and license information, please view
- * the file license.txt that was distributed with this source code.
- */
-
-namespace Nette\Reflection;
-
-use Nette;
-
-
-
-/**
- * Basic annotation implementation.
- *
- * @author     David Grudl
- */
-class Annotation extends Nette\Object implements IAnnotation
-{
-
-	public function __construct(array $values)
-	{
-		foreach ($values as $k => $v) {
-			$this->$k = $v;
-		}
-	}
-
-
-
-	/**
-	 * Returns default annotation.
-	 * @return string
-	 */
-	public function __toString()
-	{
-		return $this->value;
-	}
-
-}
